Makes downloads controller global
