+++ app/views/prior_authority/steps/primary_quote_summary/_service_cost_card.html.erb
@@ -25,7 +25,7 @@
       <div class="govuk-summary-list__row">
         <dt class="govuk-summary-list__key"><%= t('.service_quote_upload') %></dt>
         <dd class="govuk-summary-list__value">
-          <%= govuk_link_to(quote.document.file_name , prior_authority_download_path(quote.document)) if quote.document %>
+          <%= govuk_link_to(quote.document.file_name , download_path(quote.document)) if quote.document %>
         </dd>
       </div>
       <div class="govuk-summary-list__row">
